Cut the comments, split the tests, keep the digest daily
I'd built a habit of answering review questions in the code: 85 comment
lines across these files, most of them restating the line below or
arguing for a decision. That belongs in a commit message, where it's
attached to the change rather than to the file forever. Fizzy runs at
1.6% across all its models and controllers, and card.rb, comment.rb,
closeable.rb, closures_controller.rb and card_scoped.rb have none at
all. AGENTS.md now says so, with the four kinds never to write.

Two tests inserted 25 and 35 records to prove Pagy paginates. Deleted
both. The one that remains covers our own logic — page two renders its
own frame and no second form — and builds the fewest records that
produce a page two. AGENTS.md picks up the rule, along with system
tests being a last resort.

App::CommentsControllerTest had grown to 32 tests over four
controllers, because the tests never followed the controller split.
Now four files that mirror them. Three separate return_path tests
became one, and the two nav-highlighting tests moved to
app_helper_test where they exercise the helper directly rather than
through full integration requests.

The digest goes back to daily: a moderation queue shouldn't nag, and
six emails on the day a post takes off is the worst possible time. It
now reports the whole queue rather than just the new arrivals, so the
count stops claiming a backlog is smaller than it is, while still only
sending when something new has come in.

CommentMailer never wrapped in I18n.with_locale, so eleven locales of
translation were dead and every digest went out in English.

Brakeman flagged the comment author link. Post::Comment#link_url
returns nil unless URI.parse gives an http or https scheme, verified
against case variants, whitespace, embedded newlines, data:, vbscript:
and protocol-relative URLs. Ignored with a note.
