+++ app/controllers/app/comments/closures_controller.rb
@@ -1,11 +1,8 @@
-# Creating a closure closes comments on the post; destroying it reopens them.
 class App::Comments::ClosuresController < AppController
   include CommentModeration
 
   before_action :load_comment
 
-  # Closing removes nothing, so both actions return to the comment you were
-  # reading rather than out to the list.
   def create
     @comment.post.close_comments!
     redirect_to comment_path, notice: "Comments closed."
