+++ app/controllers/app/comments/replies_controller.rb
@@ -1,5 +1,3 @@
-# Replying to a comment that's already approved. Approving and replying at once
-# is App::Comments::ApprovalsController's job.
 class App::Comments::RepliesController < AppController
   include CommentModeration
 
@@ -15,8 +13,6 @@ class App::Comments::RepliesController < AppController
     end
   end
 
-  # Deleting your reply frees you to write another, so go back to the comment
-  # rather than all the way out to the list.
   def destroy
     @comment.replies.find(params[:id]).destroy!
 
@@ -25,7 +21,6 @@ class App::Comments::RepliesController < AppController
 
   private
 
-    # Only an approved top-level comment can be replied to.
     def load_comment
       @comment = @blog.comments.approved.top_level.includes(:replies).find(params[:comment_id])
     end
