+++ app/jobs/check_post_comment_job.rb
@@ -5,8 +5,6 @@ class CheckPostCommentJob < ApplicationJob
     comment = Post::Comment.find_by(id: comment_id)
     return unless comment
 
-    # Clean comments stay pending until the blogger approves them. Spam never
-    # reaches the moderation queue at all.
     if comment.spam?
       Rails.logger.info("[CheckPostCommentJob] Spam detected, destroying comment #{comment.id}")
       comment.destroy!
