+++ app/mailers/comment_mailer.rb
@@ -5,9 +5,11 @@ class CommentMailer < MailpaceMailer
     @blog = params[:blog]
     @comments = params[:comments]
 
-    mail(
-      to: @blog.user.email,
-      subject: t("comments.mailer.subject")
-    )
+    I18n.with_locale(@blog.locale) do
+      mail(
+        to: @blog.user.email,
+        subject: t("comments.mailer.subject")
+      )
+    end
   end
 end
