+++ app/views/app/comments/_blank_slate.html.erb
@@ -1,5 +1,4 @@
 <div class="rounded-xl border border-slate-200 bg-slate-50 p-6 sm:p-8 dark:border-slate-700 dark:bg-slate-800/50">
-  <%# On a post page the header above is already doing the h1's job %>
   <% if @post %>
     <h2 class="text-2xl font-semibold tracking-tight text-slate-900 dark:text-slate-50">No comments yet</h2>
     <p class="mt-2 leading-7 text-slate-600 dark:text-slate-300">
