+++ app/views/app/posts/_posts_list.html.erb
@@ -15,7 +15,6 @@
             </div>
             <%= pill @drafts.size %>
           </div>
-          <%# Only here when there's no Published section to carry it %>
           <%= render "info_toggle" if @posts.empty? %>
         </div>
